refactor: move checkValidRev and isInt into separate file
